Busqueda aleatoria de funciones de uno o todos los paquetes instalados.

Esta funcion permite obtener una funcion aleatoria de alguno de los paquete instalados con fines de aprendizaje. Es decir, la idea es seleccionar al azar una funcion y la funcion que salga sorteada es con la que se debe experimentar para aprender sobre ella.

rfunction(package = "base", ALL = FALSE, library = TRUE)

Arguments

package
por defecto en la funcion esta cargado el paquete base. Pero se puede seleccionar cualquiera de los paquetes que fueron instalados en su computadora.

ALL
por defecto es FALSE. Si es TRUE significa que la busqueda aleatoria se va a buscar entre todos los paquetes que estan cargados en la sesion de trabajo actual. Si adicionalmente se selecciono un paquete en package este sera incorporado al muestreo.

library
por defecto es TRUE. Esto significa que el paquete seleccionado queda cargado en la rurta de busqued de nombres de R. Caso contrario el paquete no va a quedar cargado.

Source

http://bit.ly/Instructivo_package_IntRo

Value

Esta funcion va a imprimir en pantalla la funcion sorteada, el paquete al que pertenece y la cantidad de funciones que ese paquqte tiene. Adicionalmente se va imprimir una descripcion del paquete y se va a abrir un device de la ayuda de la funcion sorteada.

Details

Si se selecciona un paquete que no esta cargado en la sesion de trabajo no se afecta la ruta de busqueda de nombres de R. El paquete seleccionado no va a quedar cargado en la ruta actual unicamente si selecciona library = FALSE.Si se selecciona un paquete que no esta instalado la funcion le va a preguntar si desea instalarlo mediante una ventana emergente. Si no desea instalar o la instalacion es fallida por algun motivo la funcion va a debolver un muestreo del paquete 'base'.

Examples

rfunction()

#> 
#>  - The function of the day is: warningCondition 
#>  - Package: base --> 1232 functions 
#>  
#>  - Package Description: Base R functions.

rfunction(ALL = TRUE)

#> 
#>  - The function of the day is: findRestart 
#>  - Package: base --> 1232 functions 
#>  
#>  - Package Description: Base R functions.

rfunction(package = "stats")

#> 
#>  - The function of the day is: contr.sum 
#>  - Package: stats --> 448 functions 
#>  
#>  - Package Description: R statistical functions.